+++ src/main/java/org/example/prumpt_be/dto/entity/Category.java
@@ -1,4 +1,4 @@
-package org.example.prumpt_be.entity;
+package org.example.prumpt_be.dto.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -8,15 +8,10 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class Tag {
+public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     private String name;
-
-    //name만 받는 생성자 추가
-    public Tag(String name) {
-        this.name = name;
-    }
 }
