+++ src/main/java/org/example/prumpt_be/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package org.example.prumpt_be.repository;
 
-import org.example.prumpt_be.entity.Category;
+import org.example.prumpt_be.dto.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
