+++ src/main/java/org/example/prumpt_be/repository/ModelCategoryRepository.java
@@ -1,6 +1,6 @@
 package org.example.prumpt_be.repository;
 
-import org.example.prumpt_be.entity.ModelCategory;
+import org.example.prumpt_be.dto.entity.ModelCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ModelCategoryRepository extends JpaRepository<ModelCategory, Long> {
