+++ src/main/java/org/example/prumpt_be/repository/PromptClassificationRepository.java
@@ -1,7 +1,7 @@
 package org.example.prumpt_be.repository;
 
-import org.example.prumpt_be.entity.PromptClassification;
-import org.example.prumpt_be.entity.PromptClassificationId;
+import org.example.prumpt_be.dto.entity.PromptClassification;
+import org.example.prumpt_be.dto.entity.PromptClassificationId;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PromptClassificationRepository extends JpaRepository<PromptClassification, PromptClassificationId> {
