+++ src/main/java/org/example/prumpt_be/repository/PromptRepository.java
@@ -1,6 +1,6 @@
 package org.example.prumpt_be.repository;
 
-import org.example.prumpt_be.entity.Prompt;
+import org.example.prumpt_be.dto.entity.Prompt;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
