+++ src/main/java/org/example/prumpt_be/repository/PromptReviewRepository.java
@@ -1,6 +1,6 @@
 package org.example.prumpt_be.repository;
 
-import org.example.prumpt_be.entity.PromptReview;
+import org.example.prumpt_be.dto.entity.PromptReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
