+++ src/main/java/org/example/prumpt_be/repository/TagRepository.java
@@ -1,6 +1,6 @@
 package org.example.prumpt_be.repository;
 
-import org.example.prumpt_be.entity.Tag;
+import org.example.prumpt_be.dto.entity.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
