+++ src/main/java/org/example/prumpt_be/repository/TypeCategoryRepository.java
@@ -1,6 +1,6 @@
 package org.example.prumpt_be.repository;
 
-import org.example.prumpt_be.entity.TypeCategory;
+import org.example.prumpt_be.dto.entity.TypeCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TypeCategoryRepository extends JpaRepository<TypeCategory, Long> {
