+++ src/main/java/org/example/prumpt_be/service/PromptService.java
@@ -3,9 +3,9 @@ package org.example.prumpt_be.service;
 import lombok.RequiredArgsConstructor;
 import org.example.prumpt_be.dto.PromptDetailDTO;
 import org.example.prumpt_be.dto.ReviewDTO;
+import org.example.prumpt_be.dto.entity.*;
 import org.example.prumpt_be.dto.request.PromptCreateRequestDTO;
 import org.example.prumpt_be.dto.request.PromptUpdateRequestDTO;
-import org.example.prumpt_be.entity.*;
 import org.example.prumpt_be.repository.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
